+++ shaders/SimpleShader.frag
@@ -7,7 +7,7 @@ layout (location = 0) out vec4 outColor;
 layout(push_constant) uniform PushConstant
 {
     mat4 transform;
-    vec3 color;
+    mat4 normalMatrix;
 } pushConstant;
 
 void main()
