+++ src/Core/Components/RigidBody2dComponent.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <glm/gtc/matrix_transform.hpp>
+
+namespace Liara::Core::Component
+{
+    struct RigidBody2dComponent
+    {
+        glm::vec2 velocity;
+        float mass{1.0f};
+    };
+}
