+++ src/Graphics/Liara_Model.cpp
@@ -129,7 +129,9 @@ namespace Liara::Graphics
     {
         return{
             {0, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(Vertex, position)},
-            {1, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(Vertex, color)}
+            {1, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(Vertex, color)},
+            {2, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(Vertex, normal)},
+            {3, 0, VK_FORMAT_R32G32_SFLOAT, offsetof(Vertex, uv)}
         };
     }
